+++ src/main/java/Hospital/Staff/User.java
@@ -12,12 +12,11 @@ import lombok.Setter;
 
 public class User {
 
-    private Long id;
+    private Integer id;
     private String firstName;
     private String lastName;
     private String username;
     private String password;
     private String email;
-
 
 }
